backend_dmm/requirements.txt
Django==2.0.2
django-phonenumber-field==2.0.0
djangorestframework==3.7.7
flake8==3.5.0
